Grouping all of the actions on Log Objects#2096
