+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/ObjectMenuItems.tsx
@@ -1,15 +1,19 @@
 import { Typography } from "@equinor/eds-core-react";
 import { Divider, MenuItem } from "@material-ui/core";
 import React from "react";
+import { v4 as uuid } from "uuid";
 import { DispatchNavigation } from "../../contexts/navigationAction";
 import { NavigationState } from "../../contexts/navigationContext";
 import { DispatchOperation } from "../../contexts/operationStateReducer";
+import { OpenInQueryView } from "../../hooks/useOpenInQueryView";
+import LogObject from "../../models/logObject";
 import ObjectOnWellbore from "../../models/objectOnWellbore";
 import { ObjectType } from "../../models/objectType";
 import { Server } from "../../models/server";
 import Wellbore from "../../models/wellbore";
 import { colors } from "../../styles/Colors";
-import { StyledIcon, menuItemText, onClickDeleteObjects, onClickRefreshObject } from "./ContextMenuUtils";
+import { ObjectTypeToTemplateObject, StoreFunction } from "../ContentViews/QueryViewUtils";
+import { StyledIcon, menuItemText, onClickDeleteObjects, onClickRefreshObject, onClickShowGroupOnServer } from "./ContextMenuUtils";
 import { onClickCopyToServer } from "./CopyToServer";
 import { copyObjectOnWellbore, pasteObjectOnWellbore } from "./CopyUtils";
 import NestedMenuItem from "./NestedMenuItem";
@@ -26,7 +30,9 @@ export const ObjectMenuItems = (
   navigationState: NavigationState,
   dispatchOperation: DispatchOperation,
   dispatchNavigation: DispatchNavigation,
-  wellbore: Wellbore
+  openInQueryView: OpenInQueryView,
+  wellbore: Wellbore,
+  extraMenuItems: React.ReactElement[]
 ): React.ReactElement[] => {
   const objectReferences = useClipboardReferencesOfType(objectType);
   const { selectedServer, servers } = navigationState;
@@ -62,6 +68,50 @@ export const ObjectMenuItems = (
     <MenuItem key={"delete"} onClick={() => onClickDeleteObjects(dispatchOperation, checkedObjects, objectType)} disabled={checkedObjects.length === 0}>
       <StyledIcon name="deleteToTrash" color={colors.interactive.primaryResting} />
       <Typography color={"primary"}>{menuItemText("delete", objectType, checkedObjects)}</Typography>
-    </MenuItem>
+    </MenuItem>,
+    ...extraMenuItems,
+    <NestedMenuItem key={"showOnServer"} label={"Show on server"} disabled={checkedObjects.length !== 1}>
+      {servers.map((server: Server) => (
+        <MenuItem key={server.name} onClick={() => onClickShowGroupOnServer(dispatchOperation, server, wellbore, objectType, (checkedObjects[0] as LogObject)?.indexType)}>
+          <Typography color={"primary"}>{server.name}</Typography>
+        </MenuItem>
+      ))}
+    </NestedMenuItem>,
+    <NestedMenuItem key={"queryItems"} label={"Query"} icon="textField">
+      {[
+        <MenuItem
+          key={"openInQueryView"}
+          disabled={checkedObjects.length != 1}
+          onClick={() =>
+            openInQueryView({
+              templateObject: ObjectTypeToTemplateObject[objectType],
+              storeFunction: StoreFunction.GetFromStore,
+              wellUid: wellbore.wellUid,
+              wellboreUid: wellbore.uid,
+              objectUid: checkedObjects[0].uid
+            })
+          }
+        >
+          <StyledIcon name="textField" color={colors.interactive.primaryResting} />
+          <Typography color={"primary"}>Open in query view</Typography>
+        </MenuItem>,
+        <MenuItem
+          key={"newObject"}
+          disabled={checkedObjects.length != 1}
+          onClick={() =>
+            openInQueryView({
+              templateObject: ObjectTypeToTemplateObject[objectType],
+              storeFunction: StoreFunction.AddToStore,
+              wellUid: wellbore.wellUid,
+              wellboreUid: wellbore.uid,
+              objectUid: uuid()
+            })
+          }
+        >
+          <StyledIcon name="add" color={colors.interactive.primaryResting} />
+          <Typography color={"primary"}>{`New ${objectType}`}</Typography>
+        </MenuItem>
+      ]}
+    </NestedMenuItem>
   ];
 };
